[Impeller] move TrackedObjectsVK to separate file.
