+++ impeller/renderer/backend/vulkan/command_encoder_vk.cc
@@ -11,92 +11,9 @@
 #include "impeller/renderer/backend/vulkan/fence_waiter_vk.h"
 #include "impeller/renderer/backend/vulkan/gpu_tracer_vk.h"
 #include "impeller/renderer/backend/vulkan/texture_vk.h"
+#include "impeller/renderer/backend/vulkan/tracked_objects_vk.h"
 
 namespace impeller {
-
-class TrackedObjectsVK {
- public:
-  explicit TrackedObjectsVK(const std::weak_ptr<const ContextVK>& context,
-                            const std::shared_ptr<CommandPoolVK>& pool,
-                            std::unique_ptr<GPUProbe> probe)
-      : desc_pool_(context), probe_(std::move(probe)) {
-    if (!pool) {
-      return;
-    }
-    auto buffer = pool->CreateCommandBuffer();
-    if (!buffer) {
-      return;
-    }
-    pool_ = pool;
-    buffer_ = std::move(buffer);
-    is_valid_ = true;
-  }
-
-  ~TrackedObjectsVK() {
-    if (!buffer_) {
-      return;
-    }
-    pool_->CollectCommandBuffer(std::move(buffer_));
-  }
-
-  bool IsValid() const { return is_valid_; }
-
-  void Track(std::shared_ptr<SharedObjectVK> object) {
-    if (!object) {
-      return;
-    }
-    tracked_objects_.insert(std::move(object));
-  }
-
-  void Track(std::shared_ptr<const Buffer> buffer) {
-    if (!buffer) {
-      return;
-    }
-    tracked_buffers_.insert(std::move(buffer));
-  }
-
-  bool IsTracking(const std::shared_ptr<const Buffer>& buffer) const {
-    if (!buffer) {
-      return false;
-    }
-    return tracked_buffers_.find(buffer) != tracked_buffers_.end();
-  }
-
-  void Track(std::shared_ptr<const TextureSourceVK> texture) {
-    if (!texture) {
-      return;
-    }
-    tracked_textures_.insert(std::move(texture));
-  }
-
-  bool IsTracking(const std::shared_ptr<const TextureSourceVK>& texture) const {
-    if (!texture) {
-      return false;
-    }
-    return tracked_textures_.find(texture) != tracked_textures_.end();
-  }
-
-  vk::CommandBuffer GetCommandBuffer() const { return *buffer_; }
-
-  DescriptorPoolVK& GetDescriptorPool() { return desc_pool_; }
-
-  GPUProbe& GetGPUProbe() const { return *probe_.get(); }
-
- private:
-  DescriptorPoolVK desc_pool_;
-  // `shared_ptr` since command buffers have a link to the command pool.
-  std::shared_ptr<CommandPoolVK> pool_;
-  vk::UniqueCommandBuffer buffer_;
-  std::set<std::shared_ptr<SharedObjectVK>> tracked_objects_;
-  std::set<std::shared_ptr<const Buffer>> tracked_buffers_;
-  std::set<std::shared_ptr<const TextureSourceVK>> tracked_textures_;
-  std::unique_ptr<GPUProbe> probe_;
-  bool is_valid_ = false;
-
-  TrackedObjectsVK(const TrackedObjectsVK&) = delete;
-
-  TrackedObjectsVK& operator=(const TrackedObjectsVK&) = delete;
-};
 
 CommandEncoderFactoryVK::CommandEncoderFactoryVK(
     const std::weak_ptr<const ContextVK>& context)
